+++ sireum-distro/src/main/scala/SireumDistro.scala
@@ -50,7 +50,6 @@ object SireumDistro extends App {
     def setProperty(k : Object, v : Object) : Object
   }
 
-  final val METADATA_DIR = ".metadata/"
   final val BUILD_FILENAME = "BUILD"
   final val INSTALLED_FEATURES_FILENAME = "installed-features.txt"
   final val SAPP_EXT = ".sapp"
@@ -111,6 +110,11 @@ object SireumDistro extends App {
   }
 
   val sireumDir = new File(args(0))
+  val appsDir = new File(sireumDir, "apps")
+  val platformDir = new File(appsDir, "platform")
+  val metadataDir = new File(sireumDir, ".metadata")
+  val metadataAppsDir = new File(metadataDir, "apps")
+  val metadataPlatformDir = new File(metadataAppsDir, "platform")
   var unmanagedDir = new File(System.getProperty("user.home"))
 
   var _features = Map[String, List[String]]()
@@ -180,16 +184,6 @@ object SireumDistro extends App {
   }
 
   try {
-    if ((util.Properties.javaVersion.charAt(2) - '0') < 7) {
-      outPrintln("This version of Sireum requires at least Java 7")
-      deleteJar
-      sys.exit(-1)
-    }
-    if (util.Properties.versionString.substring(8, 12) != "2.11") {
-      outPrintln("This version of Sireum requires Scala 2.11")
-      deleteJar
-      sys.exit(-1)
-    }
     if (OS_STRING == "unsupported") {
       outPrintln(
         "Running on an unsupported platform: some features maybe unavailable")
@@ -266,16 +260,17 @@ object SireumDistro extends App {
           if (last) {
             errPrintln("Please specify features to uninstall")
           } else {
-            updateScriptAndPlatform
             var allFound = false
             for (j <- i + 1 until args.length if !allFound) {
               if (args(j) == "all")
                 allFound = true
             }
             if (allFound) uninstall("all")
-            else
+            else {
+              updateScriptAndPlatform
               for (j <- i + 1 until args.length)
                 uninstall(args(j))
+            }
           }
         case "update" =>
           if (last) {
@@ -385,7 +380,7 @@ object SireumDistro extends App {
     Class.forName(cliClassName).newInstance.asInstanceOf[Cli]
   }
 
-  def readBuild = readLine(new File(sireumDir, METADATA_DIR + BUILD_FILENAME))
+  def readBuild = readLine(new File(metadataDir, BUILD_FILENAME))
 
   def readLine(file : File) = {
     val r = new BufferedReader(new FileReader(file))
@@ -433,6 +428,7 @@ object SireumDistro extends App {
             status = delete(f, false) && status
         }
       }
+      saveInstalledFeatures(List("Platform.sapp"))
       ProgressPrinter.last
       if (status) {
         outPrintln("... done!")
@@ -481,8 +477,7 @@ object SireumDistro extends App {
               val featureDir = new File(sireumDir, featureRelPath)
               if (featureDir.exists)
                 featureDir.delete
-              val mFeatureDir = new File(sireumDir, METADATA_DIR +
-                featureRelPath)
+              val mFeatureDir = new File(metadataDir, featureRelPath)
               if (mFeatureDir.exists)
                 delete(mFeatureDir, false)
             }
@@ -555,8 +550,7 @@ object SireumDistro extends App {
 
   def isDownloaded(f : File) =
     if (f.getName.endsWith(SAPP_EXT))
-      new File(sireumDir, METADATA_DIR + relativize(sireumDir, f) +
-        CHECKSUM_SUFFIX).exists
+      new File(metadataDir, relativize(sireumDir, f) + CHECKSUM_SUFFIX).exists
     else
       f.exists
 
@@ -569,7 +563,7 @@ object SireumDistro extends App {
     val checksum = checksums(scriptName)
 
       def getPlatformFileUpdates : List[(String, String)] = {
-        val platformDir = new File(sireumDir, METADATA_DIR + "apps/platform")
+        val platformDir = metadataPlatformDir
         var result = List[(String, String)]()
         if (platformDir.isDirectory) {
           for (f <- platformDir.listFiles) {
@@ -671,9 +665,9 @@ object SireumDistro extends App {
               if (!features.contains(fPath)) {
                 if (!isAppFile(file))
                   update(fPath, file, None, true)
-                else if (isApp && !new File(sireumDir, METADATA_DIR + fPath +
+                else if (isApp && !new File(metadataDir, fPath +
                   CHECKSUM_SUFFIX).exists &&
-                  (file.getParentFile != new File(sireumDir, "apps")))
+                  (file.getParentFile != appsDir))
                   if (downloadFile(false, fPath, file))
                     downloadCount += 1
               } else if (features.contains(fPath) &&
@@ -700,7 +694,7 @@ object SireumDistro extends App {
       }
 
       def updateExisting(dirFile : File, relDirPath : String) {
-        val metaPath = new File(sireumDir, METADATA_DIR).getAbsolutePath
+        val metaPath = metadataDir.getAbsolutePath
         for (f <- dirFile.listFiles)
           if (f.isDirectory) {
             if (dirFile.getAbsolutePath != metaPath)
@@ -712,16 +706,13 @@ object SireumDistro extends App {
       }
 
     if (isApp) {
-      val metaDir = new File(sireumDir, METADATA_DIR)
-      val metaAppDir = new File(metaDir, "apps")
-
-      if (metaAppDir.exists)
-        for (d <- metaAppDir.listFiles) {
+      if (metadataAppsDir.exists)
+        for (d <- metadataAppsDir.listFiles) {
           if (d.isDirectory)
             for (f <- d.listFiles) {
               if (f.getName.endsWith(CHECKSUM_SUFFIX)) {
                 val currentChecksum = readLine(f)
-                var filePath = relativize(metaDir, f)
+                var filePath = relativize(metadataDir, f)
                 filePath = filePath.substring(0, filePath.length -
                   CHECKSUM_SUFFIX.length)
                 update(filePath, new File(sireumDir, filePath),
@@ -765,14 +756,14 @@ object SireumDistro extends App {
   }
 
   def downloadBuild {
-    val buildFile = new File(sireumDir, METADATA_DIR + BUILD_FILENAME)
+    val buildFile = new File(metadataDir, BUILD_FILENAME)
     downloadFile(buildFile.exists, BUILD_FILENAME, buildFile)
   }
 
   def deleteRemainingAppFiles(filePath : String) {
     val appPath = getAppPath(filePath)
     val dir = new File(sireumDir, appPath).getParentFile
-    val f = new File(sireumDir, METADATA_DIR + filePath + ".filelist")
+    val f = new File(metadataDir, filePath + ".filelist")
     if (!f.exists)
       return
     val r = new LineNumberReader(new FileReader(f))
@@ -921,9 +912,7 @@ object SireumDistro extends App {
     file.getAbsolutePath.substring(baseFile.getAbsolutePath.length + 1).
       replace("\\", "/")
 
-  def isManaged(file : File) =
-    file.getParentFile.getAbsolutePath != new File(sireumDir, "apps").
-      getAbsolutePath
+  def isManaged(file : File) = file.getParentFile != appsDir
 
   def installApp(file : File) {
     val relPath = relativize(sireumDir, file)
@@ -944,9 +933,9 @@ object SireumDistro extends App {
       val fileList =
         if (managed) {
           val checksums = getChecksums
-          write(new File(sireumDir, METADATA_DIR + relPath + CHECKSUM_SUFFIX),
+          write(new File(metadataDir, relPath + CHECKSUM_SUFFIX),
             checksums(relPath))
-          Some(new PrintWriter(new FileWriter(new File(sireumDir, METADATA_DIR +
+          Some(new PrintWriter(new FileWriter(new File(metadataDir,
             relPath + ".filelist"))))
         } else
           None
@@ -1061,21 +1050,24 @@ Sireum Distro managed apps are currently running.""")
 
   def delete(file : File, onExit : Boolean) : Boolean = {
     if (file.isDirectory)
-      for (f <- file.listFiles) {
-        if (f.isDirectory) {
-          delete(f, onExit)
-        } else {
-          if (onExit) f.deleteOnExit
-          else Files.delete(f.toPath)
+      if (file == platformDir) return true
+      else
+        for (f <- file.listFiles) {
+          if (f.isDirectory) {
+            delete(f, onExit)
+          } else {
+            if (onExit) f.deleteOnExit
+            else Files.delete(f.toPath)
+          }
+          ProgressPrinter.next
         }
-        ProgressPrinter.next
-      }
     if (onExit) {
       file.deleteOnExit
       true
     } else
       try {
-        Files.delete(file.toPath)
+        if (file != appsDir)
+          Files.delete(file.toPath)
         true
       } catch {
         case e : IOException => false
@@ -1106,7 +1098,6 @@ Sireum Distro managed apps are currently running.""")
   }
 
   def cleanApps {
-    val appsDir = new File(sireumDir, "apps")
     if (appsDir.exists) {
       deleteAppsBackups(appsDir)
     }
@@ -1210,8 +1201,7 @@ Sireum Distro managed apps are currently running.""")
   }
 
   def loadInstalledFeatures = {
-    val installedFeaturesFile = new File(sireumDir, METADATA_DIR +
-      "installed-features.txt")
+    val installedFeaturesFile = new File(metadataDir, "installed-features.txt")
     if (!installedFeaturesFile.exists)
       List[String]()
     else {
@@ -1229,8 +1219,7 @@ Sireum Distro managed apps are currently running.""")
   }
 
   def saveInstalledFeatures(installedFeatures : List[String]) {
-    val installedFeaturesFile = new File(sireumDir, METADATA_DIR +
-      "installed-features.txt")
+    val installedFeaturesFile = new File(metadataDir, "installed-features.txt")
     installedFeaturesFile.getParentFile.mkdirs
     val pw = new PrintWriter(new FileWriter(installedFeaturesFile))
     for (f <- installedFeatures)
